Asset and Album to Entity
