+++ Shared/Entity/Album.swift
@@ -1,0 +1,10 @@
+import Foundation
+import Photos
+import UIKit
+
+struct Album: Identifiable {
+  var id: String { collection.localIdentifier }
+
+  let collection: PHAssetCollection
+  var firstAsset: Asset?
+}
