+++ Shared/PhotoLibrary/PhotoLibrary.swift
@@ -6,22 +6,6 @@ import os.log
 
 private let logger = Logger(subsystem: "com.photomark.log", category: "photoLibrary")
 
-final class Asset: CustomStringConvertible, Identifiable {
-  var id: String { asset.localIdentifier }
-
-  let asset: PHAsset
-  var image: UIImage?
-  var info: [AnyHashable: Any]?
-
-  init(phAsset: PHAsset) {
-    self.asset = phAsset
-  }
-
-  var description: String {
-    "asset: \(asset), image: \(String(describing: image)), info: \(String(describing: info))"
-  }
-}
-
 struct PhotoLibrary {
   enum AuthorizationAction {
     case openSettingApp
